dt: binding: Add specified-clock for discoverable rate clocks
The frequency of some fixed rate external oscillators on some SoCs (for
example TZ1090's XTAL1) are specified by the board using pull-ups and
pull-downs of configuration pins which are automatically latched on
reset and available in an SoC register, so that the boot ROM and OS can
automatically discover it.

These clocks are very similar to fixed-clocks since the rate never
changes, but we still want to be able to automatically discover the
frequency, so add a new binding to describe such clocks. Instead of
using a "clock-frequency" property, use "reg" or "syscon-reg",
"bit-mask", "bit-shift" properties to describe the register field. The
mapping from register values to clock frequencies is specified in a
"table" property which contains pairs of register values and
frequencies.

There is scope to extend the binding further in future to handle other
mapping types, by using different properties instead of "table".

Signed-off-by: James Hogan <james.hogan@imgtec.com>
Cc: Mike Turquette <mturquette@linaro.org>
Cc: Rob Herring <robh+dt@kernel.org>
Cc: Pawel Moll <pawel.moll@arm.com>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Ian Campbell <ijc+devicetree@hellion.org.uk>
Cc: Kumar Gala <galak@codeaurora.org>
Cc: devicetree@vger.kernel.org
---
Changes in v4:
- Add syscon phandle interface as alternative way to access register
  specifying clock rate.
- Add clock-accuracy property to match fixed-clock.

Changes in v3:
- Remove unused & unimplemented gpios property. Not sure why it was
  there in the first place.

I could easily have extended the fixed-rate binding instead, but there'd
be no properties in common except the standard common clock properties
so it seemed like it deserved a separate binding. I'm open to arguments
to the contrary or better compatible string names though.

Changes in v2:
- Borrow bit-mask and bit-shift bindings from Mike's mux clock binding
  proposals.