clk-fixed-rate: Support specified-clock binding
Implement the new specified-clock DT binding which uses a table to look
up the fixed frequency based on the value in a register.

The discovery is done entirely while setting up the clock, since the
clock rate is still fixed so there's no need to dynamically rediscover
the rate. The register value is obtained either via an iomap or by
obtaining a syscon regmap, then the field value is extracted and used to
look up the clock frequency in the table property.

Signed-off-by: James Hogan <james.hogan@imgtec.com>
Cc: Mike Turquette <mturquette@linaro.org>
---
Changes in v4:
- Handle syscon phandle interface as alternative way to access register
  specifying clock rate.
- handle clock-accuracy property to match fixed-clock.

No changes in v3.

Changes in v2:
- Rewrite to use a fixed clock instead of an entirely new clock type.