serial: 8250: Tolerate clock variance for max baud rate
When the UART clock is set slightly under 1.8432MHz, the 8250 driver
core doesn't permit the 115200 baud rate since it calculates the maximum
frequency to pass to uart_get_baud_rate by simply dividing the uart
clock by 16 which yields a value slightly under 115200, even though the
frequency is close enough for the UART to operate reliably.

Therefore add some tolerance in the calculation of the maximum baud
rate, specifically +1/128th (~0.8%), to allow the 115200 baud rate to
be closen for a UART clock as low as 1.8149MHz. For an external divider
set as close as possible to 1.8432MHz, this should cover every possible
input frequency above 118.9MHz.

Signed-off-by: James Hogan <james.hogan@imgtec.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Jiri Slaby <jslaby@suse.cz>
Cc: Anton Vorontsov <avorontsov@ru.mvista.com>
Cc: Peter Hurley <peter@hurleysoftware.com>
Cc: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Cc: linux-serial@vger.kernel.org
---
As far as I can tell from reading the link below, this tolerance should
be okay, and it definitely covers the range of input frequencies I
expect for my particular hardware. I'm open to better ways of handling
it though.

https://electronics.stackexchange.com/questions/5850/how-critical-are-uart-frequencies