dt: binding: Add binding for TZ1090 HEP clocks
The TZ1090 High End Peripheral (HEP) register region controls several
clocks for the HEP peripherals.

Create a clock binding for the provider of these clocks, and a header
file in <dt-bindings/clock/> for enumerating the provided clocks.

There are 3 clocks provided, the 2D block clock, the PDP/PDI clocks, and
the DDR enable clock which is fed back into the TOP clocks for dividing
down.

Signed-off-by: James Hogan <james.hogan@imgtec.com>
Cc: Mike Turquette <mturquette@linaro.org>
Cc: Rob Herring <robh+dt@kernel.org>
Cc: Pawel Moll <pawel.moll@arm.com>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Ian Campbell <ijc+devicetree@hellion.org.uk>
Cc: Kumar Gala <galak@codeaurora.org>
Cc: linux-metag@vger.kernel.org
Cc: devicetree@vger.kernel.org
---
Changes in v3:
- Rename to hep-cr and add syscon to compatible string.

Changes since v1 (patch 15):
- New patch.
- Convert explicit DT representation of clock infrastructure using
  generic bindings to several TZ1090 specific bindings representing
  groups of TZ1090 clocks.

Rename top clock binding to top-cr and syscon